@@ -18,10 +18,10 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  flutter_platform_widgets: ^0.70.2
+  flutter_platform_widgets: any
   event_bus: ^1.0.1
   path_provider: ^1.1.0
   shared_preferences: ^0.5.2
@@ -30,11 +30,11 @@ dependencies:
   image_picker: ^0.6.7+11
   camera: ^0.5.2+1
 
-  dio: ^3.0.10
-  retrofit: ^1.3.4
-  logger: ^0.9.4
-  rxdart: ^0.24.1
-  json_serializable: ^3.5.0
+  dio: any
+  retrofit: any
+  logger: any
+  rxdart: any
+  json_serializable: any
 
   flutter:
     sdk: flutter
