@@ -1,5 +1,5 @@
 import 'dart:async';
 
-extension FutureExtension on Future {
-  get stream => Stream.fromFuture(this);
-}
+// extension FutureExtension on Future {
+//   get stream => Stream.fromFuture(this);
+// }
